1. Фильтры для объявлений машин [#4745 #540569]

+++ src/main/java/ru/job4j/cars/service/post/SimplePostService.java
@@ -2,31 +2,66 @@ package ru.job4j.cars.service.post;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.job4j.cars.dto.FileDto;
+import ru.job4j.cars.model.File;
 import ru.job4j.cars.model.Post;
 import ru.job4j.cars.repository.post.PostRepository;
+import ru.job4j.cars.service.file.FileService;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
 public class SimplePostService implements PostService {
 
     private final PostRepository postRepository;
+    private final FileService fileService;
 
     @Override
-    public Post create(Post post) {
+    public Post create(Post post, List<FileDto> images) {
+        var photos = saveNewFile(images);
+        post.setPhotos(photos);
         return postRepository.create(post);
     }
 
-    @Override
-    public void update(Post post) {
-        postRepository.update(post);
+    private List<File> saveNewFile(List<FileDto> images) {
+        return images.stream()
+                .map(fileService::save)
+                .toList();
     }
 
     @Override
-    public void delete(int postId) {
-        postRepository.delete(postId);
+    public void update(Post post, List<FileDto> images) {
+        boolean isAllNewFilesEmpty = images.stream()
+                .allMatch(image -> image.getContent().length == 0);
+        if (isAllNewFilesEmpty) {
+            postRepository.update(post);
+            return;
+        }
+        var newPhotos = saveNewFile(images);
+        deleteFile(post);
+        post.setPhotos(newPhotos);
+        postRepository.update(post);
+    }
+
+    private void deleteFile(Post post) {
+        List<Integer> fileIds = post.getPhotos().stream()
+                .map(File::getId)
+                .toList();
+        fileService.deleteAllByIds(fileIds);
+    }
+
+    @Override
+    public boolean delete(int postId) {
+        var postOptional = findById(postId);
+        boolean isDeleted = false;
+        if (postOptional.isPresent()) {
+            isDeleted = postRepository.delete(postId);
+            deleteFile(postOptional.get());
+        }
+        return isDeleted;
     }
 
     @Override
@@ -37,5 +72,20 @@ public class SimplePostService implements PostService {
     @Override
     public Optional<Post> findById(int postId) {
         return postRepository.findById(postId);
+    }
+
+    @Override
+    public List<Post> findByCreatedAfter() {
+        return postRepository.findByCreatedAfter();
+    }
+
+    @Override
+    public List<Post> findAllWithPhotos() {
+        return postRepository.findAllWithPhotos();
+    }
+
+    @Override
+    public List<Post> findByLikeCarBrand(String key) {
+        return postRepository.findByLikeCarBrand(key);
     }
 }
